feat: Add statusToken endpoint to PurchaseController
The code changes in the `PurchaseController` add a new `statusToken` endpoint. This endpoint handles the status of a transaction token received from Transbank. It checks if the token is present and then confirms the transaction status. If the token is missing, it returns a 400 error. If there is an error during the confirmation process, it returns a 500 error. Otherwise, it returns the transaction status.

This commit adds the `statusToken` endpoint to the `PurchaseController`.